使用EventBus， 在父activity中控制每个fragment的， 切换recyclerView的布局样式，

+++ app/src/main/java/com/aya/bicycle006/model/D.java
@@ -160,7 +160,7 @@ public class D {
 
     public static class RatingEntity {
         private int max;
-        private double average;
+        private float average;
         private String stars;
         private int min;
 
@@ -172,11 +172,11 @@ public class D {
             this.max = max;
         }
 
-        public double getAverage() {
+        public float getAverage() {
             return average;
         }
 
-        public void setAverage(double average) {
+        public void setAverage(float average) {
             this.average = average;
         }
